Integrate Data Loader with CORE
