+++ core/archtype.go
@@ -1,4 +1,4 @@
-package main
+package core
 
 import (
 	"encoding/json"
@@ -8,24 +8,13 @@ import (
 	bolt "go.etcd.io/bbolt"
 )
 
-type Archetype struct {
-	Name        string             `json:"name"`
-	Description string             `json:"description"`
-	Attributes  map[string]float64 `json:"Attributes"`
-	Abilities   map[string]float64 `json:"Abilities"`
-}
-
-type ArchetypesData struct {
-	Archetypes map[string]Archetype `json:"archetypes"`
-}
-
-func archDisplay(archetypes *ArchetypesData) {
+func DisplayArchetypes(archetypes *ArchetypesData) {
 	for key, archetype := range archetypes.Archetypes {
 		fmt.Println(key, archetype)
 	}
 }
 
-func archLoadJSON(fileName string) (*ArchetypesData, error) {
+func LoadArchetypesFromJSON(fileName string) (*ArchetypesData, error) {
 	file, err := os.ReadFile(fileName)
 	if err != nil {
 		return nil, err
@@ -37,7 +26,6 @@ func archLoadJSON(fileName string) (*ArchetypesData, error) {
 		return nil, err
 	}
 
-	// Iterate over the loaded archetypes and print a line for each.
 	for key, archetype := range data.Archetypes {
 		fmt.Printf("Loaded archetype '%s': %s - %s\n", key, archetype.Name, archetype.Description)
 	}
@@ -45,14 +33,8 @@ func archLoadJSON(fileName string) (*ArchetypesData, error) {
 	return &data, nil
 }
 
-func archWriteBolt(archetypes *ArchetypesData, dbPath string) error {
-	db, err := bolt.Open(dbPath, 0600, nil)
-	if err != nil {
-		return err
-	}
-	defer db.Close()
-
-	return db.Update(func(tx *bolt.Tx) error {
+func StoreArchetypes(kp *KeyPair, archetypes *ArchetypesData) error {
+	return kp.db.Update(func(tx *bolt.Tx) error {
 		bucket, err := tx.CreateBucketIfNotExists([]byte("Archetypes"))
 		if err != nil {
 			return err
@@ -72,16 +54,10 @@ func archWriteBolt(archetypes *ArchetypesData, dbPath string) error {
 	})
 }
 
-func archLoadBolt(dbPath string) (*ArchetypesData, error) {
-	db, err := bolt.Open(dbPath, 0600, &bolt.Options{ReadOnly: true})
-	if err != nil {
-		return nil, err
-	}
-	defer db.Close()
-
+func LoadArchetypes(kp *KeyPair) (*ArchetypesData, error) {
 	archetypesData := &ArchetypesData{Archetypes: make(map[string]Archetype)}
 
-	err = db.View(func(tx *bolt.Tx) error {
+	err := kp.db.View(func(tx *bolt.Tx) error {
 		bucket := tx.Bucket([]byte("Archetypes"))
 		if bucket == nil {
 			return fmt.Errorf("archetypes bucket does not exist")
@@ -103,4 +79,12 @@ func archLoadBolt(dbPath string) (*ArchetypesData, error) {
 	}
 
 	return archetypesData, nil
+}
+
+func (k *KeyPair) StoreArchetypes(archetypes *ArchetypesData) error {
+	return StoreArchetypes(k, archetypes)
+}
+
+func (k *KeyPair) LoadArchetypes() (*ArchetypesData, error) {
+	return LoadArchetypes(k)
 }
